Add files via upload
FizzBuzz is a popular game among the British kids. In this game, each time when a number that is divisible by 3 comes they have to say Fizz, if a number that is divisible by 5 comes they have to say Buzz and if a number is divisible by both 3 and 5 they have to say FizzBuzz. 👥👥